requirements.txt
numpy>=1.24.0
pandas>=2.0.0
langchain>=0.1.0
langchain-community>=0.0.10
langchain-core>=0.1.0
langchain-google-genai>=0.0.5
langgraph>=0.0.15
faiss-cpu>=1.7.4
python-dotenv>=1.0.0
pydantic>=2.0.0
tenacity>=8.0.0
streamlit>=1.22.0